[MS] Add useFileActions composable to isolate file action functions
